del main test_metrics.py file
